requirements.txt
#apache-airflow==2.7.1
pendulum==3.0.0 # Raises 'module' object not callable error2.1.0
apache-airflow==2.9.0